adehabitat-package {adehabitat}	R Documentation

adehabitat: a Package for the Analysis of the Space Use by Animals

Description

The package adehabitat has been designed to allow the analysis of the space use by animals. This package is organised in four main parts: (i) management of raster maps, (ii) habitat selection/ecological niche analysis, (iii) home range estimation and (iv) analysis of animals trajectories. These four parts interact with each other to facilitate the analysis. Note that adehabitat strongly relies on the package ade4, which provides numerous functions for the analysis of multivariate data.

Details

Package: 	 adehabitat

Type: 	 Package

Version: 	 1.8.3

Date: 	 2009-05-04

License: 	 GPL version 2 or newer

The four parts of the packages are described more in detail below:

Management of raster mapsHowever, it provides numerous functions allowing to import and export raster maps from/to Geographic information system, to compute buffers around points or lines, to identify the value of environmental variables at given spatial locations, to count the number of points of a pattern in each pixel of a map, etc. Two basic object classes allow to manage ratser maps with adehabitat: the class asc is intended to store basic raster map (see help(import.asc)), whereas the class kasc is intended to store multi-layer maps (all covering the same area with the same resolution, see help(as.kasc)). For additional information this part of the package, see the tutorial available in the package. Type demo(rastermaps) for a demonstration of the package capabilities. Note that the package sp also provides many interesting functions to manage raster maps, and adehabitat provide functions of conversion to the classes of the package (see help(kasc2spixdf)).

Habitat selection/ecological niche analysismany methods have been included in the package to render statistical methods allowing the analysis of habitat selection available to wildlife ecologists. Many of them are factor analyses of the niche or of the habitat selection, but other methods are also available. They include the selection ratios (see help(wi)), the Ecological niche factor analysis (see help(enfa)), the Mahalanobis distances (see help(mahasuhab)) and their factorial decomposition (the MADIFA, see help(madifa)) or the algorithm DOMAIN (see help(domain)). Other common methods, such as the resource selection functions can also be used with the rest of the R environment. Note that the package also include functions allowing the analysis of habitat selection using radio-tracking data, such as the compositional analysis (see help(compana)), the eigenanalysis of selection ratios (see help(eisera)) or the K-select analysis (see help(kselect)). An overview of these methods is available by typing demo(nichehs).

home range estimationmany methods are available to estimate the home range of animals relocated using radio-tracking data. Common methods of estimation are available, such as the Minimum convex polygon (see help(mcp)), the kernel estimation of the utilization distribution (see help(kernelUD)), the cluster home range (see help(clusthr)) or the nearest neighbour convex hull (see help(NNCH)). Note that Paolo Cavallini has designed a website dedicated to the analysis of space use by animals, which contain a wiki page, a tutorial for the home range estimation using R and adehabitat and a forum (URL: http://www.faunalia.it/animov/index.php). Several methods of these part of the package have been included following discussions that arose on this forum (especially, the nearest neighbour convex hull and the brownian bridge kernel). Type demo(homerange) for examples of use of these functions.

The analysis of animals' trajectoriesThis part is the most recent one in the package. A new class designed to store animals' trajectories has been included in the package, the class ltraj (see help(as.ltraj)). Two types of trajectories can be handled with adehabitat: for trajectories of type I, the time is not recorded for the relocations (e.g. the sampling of the tracks of an animal in the snow). For trajectories of type II, the time has been recorded during sampling (e.g. radio-tracking, GPS, Argos monitoring). Many descriptive parameters are automatically computed (relative or turning angles, distance between successive relocations, mean squared displacement). Many functions allow the management and analysis of trajectories, through the analysis of these parameters (e.g. tests of independence, see help(wawotest,indmove), first passage time, see fpt). The rediscretisation of trajectories of type I is also possible (help(redisltraj)). Many graphical functions are available for the exploration of the trajectory properties (plot,plotltr,sliwinltr), etc. A new partitioning algorithm has been added (but it is still under research) to partition animals trajectories into segment with homogeneous properties (see modpartltraj). Further details can be found on the help page of the function as.ltraj. For a demonstration, type demo(managltraj) or demo(analysisltraj).

Author(s)

Clement Calenge, with contributions from Mathieu Basille, Stephane Dray, Manuela Royer and Scott Fortmann-Roe

Maintainer: Clement Calenge <clement.calenge@oncfs.gouv.fr>

References

Calenge, C. (2006) The package adehabitat for the R software: a tool for the analysis of space and habitat use by animals. Ecological Modelling, 197, 516-519

See Also

ade4

Examples

## Not run: 

## For examples of use of mapping capabilities
demo(rastermaps)

## For examples of use of functions for
## habitat selection and niche analysis
demo(nichehs)

## For example of home range estimation
demo(homerange)

## For example of trajectory management and analysis
demo(managltraj)
demo(analysisltraj)

## End(Not run)

[Package adehabitat version 1.8.18 Index]