albatross {adehabitat}	R Documentation

Argos Monitoring of Adult Albatross Movement

Description

This data set contains the relocations of 6 adult albatross monitored in the Crozets Islands by the team of H. Weimerskirch from the CEBC-CNRS (Centre d'Etudes Biologiques de Chize, France).

Usage

data(albatross)

Format

This data set is an object of class ltraj.

Details

The coordinates are given in meters (UTM - zone 42).

Source

http://suivi-animal.u-strasbg.fr/index.htm

Examples

data(albatross)

plot(albatross)

[Package adehabitat version 1.8.18 Index]