as.area {adehabitat}	R Documentation

Objects of Class "area"

Description

Objects of class area are used to store the information on the shape of vectorised objects.
An object of class area is a data frame with three variables. The first variable is a factor defining the polygons.
The second and third variables are the xy coordinates of the polygon vertices in the order where they are found. This kind of objects are current in the package ade4, though this package does not define area as a special class.

Usage

as.area(x)

Arguments

x
a data frame with three columns

Value

Returns an object of class area

See Also

area.plot (package ade4) for other operations on objects of class area, plot.area to display objects of this class, and area2dxf for exportation of the objects area toward a GIS.

Examples

data(elec88)
ar <- as.area(elec88$area)
plot(ar)

[Package adehabitat version 1.8.18 Index]