bighorn {adehabitat}	R Documentation

Radio-Tracking of Bighorn Sheeps

Description

This data set describes the habitat use and availability for 6 bighorn sheeps monitored by radio-tracking (Arnett et al. 1989, in Manly et al., 2003, p. 67-74). 10 habitat types are considered.

Usage

data(bighorn)

Details

The object bighorn is a list, with the following components:

used

the number of resource units used by each animal (in rows) in each habitat category (in columns).

availTrue

the availability of each habitat category.

availEstimated

a sample of available resource units in each habitat category.

References

Manly, B.F.J., McDonald, L.L., Thomas, D.L., McDonald, T.L. & Erickson, W.P. (2003) Resource selection by animals - Statistical design and Analysis for field studies. Second edition. London: Kluwer academic publishers.

[Package adehabitat version 1.8.18 Index]