burst {adehabitat}	R Documentation

ID and Bursts of an Object of Class ltraj

Description

Functions to get or set the attribute "id" or "burst" of the components of an object of class ltraj.

Usage

burst(ltraj)
burst(ltraj) <- value
id(ltraj)
id(ltraj) <- value

Arguments

ltraj
an object of class ltraj

value
a character vector of up to the same length as ltraj

Details

The functions id and burst are accessor functions, and id<- and burst<- are replacement function.

Value

For id and burst, a character vector of the same length as ltraj.

For id<- and burst<-, the updated object. (Note that the value of burst(x) <- value is that of the assignment, value, not the return value from the left-hand side.)

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

ltraj, names

Examples

data(puechcirc)
puechcirc

## To see the ID and the burst
id(puechcirc)
burst(puechcirc)

## Change the burst
burst(puechcirc) <- c("glou", "toto", "titi")
puechcirc

burst(puechcirc)[2] <- "new name"
puechcirc

## Change the ID
id(puechcirc)[id(puechcirc)=="CH93"] <- "WILD BOAR"
puechcirc

[Package adehabitat version 1.8.18 Index]