c.ltraj {adehabitat}	R Documentation

Combine Bursts of Relocations in Objects of Class "ltraj"

Description

This function combines several objects of class ltraj.

Usage

## S3 method for class 'ltraj'
c(...)

Arguments

...
objects of class ltraj to be combined

Value

An object of class ltraj.

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

ltraj for further information on the class ltraj, Extract.ltraj to extract or replace bursts of relocations, plot.ltraj and trajdyn for graphical displays, gdltraj to specify a time period

Examples

data(puechcirc)

(i <- puechcirc[1])
(j <- puechcirc[3])

(toto <- c(i,j))

[Package adehabitat version 1.8.18 Index]