capreochiz {adehabitat}	R Documentation

GPS Monitoring of one Roe Deer in Chize (France)

Description

This dataset contains the relocations of a roe deer collected using GPS collars in the Chize reserve (Deux-Sevres, France) by the ONCFS (Office national de la chasse et de la faune sauvage).

Usage

data(capreochiz)

Format

This dataset is a list containing the relocations in an object of class ltraj names locs and information on these relocations in an object named "info" (DOP, status, etc.).

Source

Sonia Said, Office national de la chasse et de la faune sauvage, CNERA-CS, 1 place Exelmans, 55000 Bar-le-Duc (France).

Examples

data(capreochiz)

plot(capreochiz$locs)

head(capreochiz$info)

[Package adehabitat version 1.8.18 Index]