chamois {adehabitat}	R Documentation

Location of Chamois Groups in the Chartreuse Mountains

Description

This data set describes the habitat use and availability by the chamois of the Chartreuse mountains (Isere, France), in 1992 and 1997. These data have been gathered during the hunting season (Fall).

Usage

data(chamois)

Details

The object chamois is a list containing the following components:
locs is a data frame containing the x and y coordinates of 198 chamois groups.
map is a map of class kasc describing the vegetation (Forest or Open areas), the distance from the ecotone Open/Forest and the slopes on the area.

References

Federation Departementale des Chasseurs de l'Isere, 65 av Jean Jaures, 38320 Eybens. France.

[Package adehabitat version 1.8.18 Index]