colasc {adehabitat}	R Documentation

Creates a Vector of Colors for a Raster Map of Type 'factor'

Description

colasc creates a vector of colors for a raster map of class asc and of type "factor".

Usage

colasc(x, ...)

Arguments

x
an object of class asc.

...
arguments named as the levels of the factor, with character values equal to the colors for these levels (see examples)

Value

Returns a character vector.

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

asc

Examples

data(puechabon)

## gets the aspect in an asc object
asp <- getkasc(puechabon$kasc, "Aspect")

## creates the vector of colors
cl <- colasc(asp, NorthEast = "blue", SouthEast = "red", 
             SouthWest = "orange", NorthWest = "green")

## graphical display
image(asp, clfac = cl, main = "Aspect", xlab = "Lambert X",
      ylab = "Lambert Y")
legend(706500, 3162000, legend=levels(asp), fill = cl, cex = 0.7)
 

[Package adehabitat version 1.8.18 Index]