domain {adehabitat}	R Documentation

Estimation of the Potential Distribution of a Species

Description

domain uses the DOMAIN algorithm to estimate the potential distribution of a species based on a list of species occurrences and on maps of the area.

Usage

domain(kasc, pts, type = c("value", "potential"), thresh = 0.95)

Arguments

kasc
an object of class kasc

pts
a data frame giving the x and y coordinates of the species occurrences.

type
a character string. The "value" of the suitability may be returned or the "potential" area of distribution

thresh
if value = "potential", a threshold value should be supplied for the suitability (by default 0.95)

Details

This function implements the DOMAIN algorithm described in Carpenter et al. (1993).

Value

Returns a matrix of class asc.

Warning

domain is restricted to maps containing only numerical variables (i.e. no factors).

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

References

Carpenter, G., Gillison, A.N. and Winter, J. (1993) DOMAIN: a flexible modelling procedure for mapping potential distributions of plants and animals. Biodiversity and conservation, 2, 667–680.

See Also

kasc for additionnal information on objects of class kasc, asc for additionnal information on matrices of class asc.

Examples

## Preparation of the data
data(puechabon)
kasc <- puechabon$kasc
kasc$Aspect <- NULL
pts <- puechabon$locs[puechabon$locs$Name == "Brock", 4:5]

## View of the data
elevation <- getkasc(kasc, "Elevation")
image(elevation)
points(pts, col = "red", pch = 16)

## Estimation of habitat suitability map
hsm <- domain(kasc, pts)
image(hsm, col = grey((1:256)/256))
contour(hsm, add = TRUE)
## Lighter areas are the most preferred areas

## Potential distribution
hsm <- domain(kasc, pts, type = "potential")
image(elevation, main = "Habitat suitability map")
image(hsm, add = TRUE, col = "orange")
points(pts, col = "red", pch = 16)

[Package adehabitat version 1.8.18 Index]