eisera {adehabitat}	R Documentation

Eigenanalysis of Selection Ratios

Description

Performs an eigenanalysis of selection ratios.

Usage

eisera(used, available, scannf = TRUE, nf = 2)
## S3 method for class 'esr'
print(x, ...)
## S3 method for class 'esr'
scatter(x, xax = 1, yax = 2, 
            csub = 1, possub = "bottomleft", ...)

Arguments

used
a data frame containing the *number* of relocations of each animal (rows) in each habitat type (columns)

available
a data frame containing the *proportion* of availability of each habitat type (columns) to each animal (rows)

scannf
logical. Whether the eigenvalues bar plot should be displayed

nf
if scannf = FALSE, an integer indicating the number of kept axes

x
an object of class esr

xax
the column number for the x-axis

yax
the column number for the y-axis

csub
a character size for the legend, used with par("cex")*csub

possub
a string of characters indicating the sub-title position ("topleft", "topright", "bottomleft", "bottomright")

...
further arguments passed to or from other methods

Details

The eigenanalysis of selection ratios has been developped to explore habitat selection by animals monitored using radio-tracking, when habitat is defined by several categories (e.g. several vegetation types, see Calenge and Dufour 2006).

This analysis can be used for both designs II (same availability for all animals, e.g. selection of the home range within the study area) and designs III (different availability, e.g. selection of the sites within the home range). In the latter case, when some available proportions are equal to zero, the selection ratios are replaced by their expectation under random habitat use, following the recommendations of Calenge and Dufour (2006).

Value

A list of class esr and dudi containing also:

available
available proportions

used
number of relocations

wij
selection ratios

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

References

Calenge, C. and Dufour, A.B. (2006) Eigenanalysis of selection ratios from animal radio-tracking data. Ecology. 87, 2349–2355.

See Also

wi for further information about the selection ratios, compana for compositional analysis.

Examples

###########################################################
###########################################################
###
###  Example given in Calenge and Dufour 2006 (design II)

data(squirrel)

## computation of the number of relocations in each habitat type
## from the data given by Aebischer et al. (1993).
## squirrel$locs give the percentage of relocations in each habitat
## type, and Aebischer et al. (1993) indicate that there are 30
## relocations per animal.
## We therefore compute the number of relocations in each habitat type
## using:
us <- round(30 * squirrel$locs / 100)

## Habitat availability 
av <- squirrel$studyarea

## Eigenanalysis of selection ratios
ii <- eisera(us, av, scannf = FALSE)

scatter(ii, grid = FALSE, clab = 0.7)

## The following graph may help the interpretation
## (see Calenge and Dufour 2006)
data(squirreloc)
locs <- squirreloc$locs
are <- squirreloc$map
co <- attr(are, "info")

li <- split(locs[,2:3], locs[,1])
opar <- par(mfrow=n2mfrow(length(li)), mar=c(0,0,2,0))
lapply(1:length(li), function(i) {
plot(are, colp = co[,2], main=names(li)[i], axes=FALSE)
points(li[[i]], pch=16, cex=1.5)
box()
})
plot(0,0, axes=FALSE, ty="n", xlim=c(-1,1), asp=1)
legend(-0.8,0.8, unique(co[,1]), fill=unique(co[,2]))
par(opar)

###########################################################
###########################################################
###
###  Example of design III

iii <- eisera(us, squirrel$mcp, scannf = FALSE)
scatter(iii, grid = FALSE, clab = 0.7)

[Package adehabitat version 1.8.18 Index]