findmaxasc {adehabitat}	R Documentation

Find Local Maxima on a Map of Class 'asc'

Description

findmaxasc finds the local maxima on a map of class asc.

Usage

findmaxasc(asc)

Arguments

asc
a map of class asc

Details

This function may be useful, among other things, to identify the local modes of the utilization distribution of an animal (e.g. estimated using kernelUD).

Value

a data frame with two columns containing the x and y coordinates of the local maxima.

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

as.asc, kernelUD

Examples

## example, to find the mode of the utilization distribution (UD) of an
## animal:

## load the data
data(puechabon)
loc <- puechabon$locs[, c("X", "Y")]
id <- puechabon$locs[, "Name"]

## Estimation of UD for the four animals
(ud <- kernelUD(loc, id))
image(ud)

## Now consider the UD of Chou
map <- ud$Chou$UD
image(map)

## Find the local maxima:
maxim <- findmaxasc(map)
points(maxim, col="red", pch=16, cex=1.5)

[Package adehabitat version 1.8.18 Index]