gdltraj {adehabitat}	R Documentation

Working with Trajectories: Specify a Time Period

Description

Gets the parts of the trajectories stored in an object of class ltraj of type II (time recorded), corresponding to a specified time period.

Usage

gdltraj(x, min, max, type = c("POSIXct", "sec", "min", "hour", "mday",
        "mon", "year", "wday", "yday")) 

Arguments

x
an object of class ltraj of type II (time recorded)

min
numeric. The beginning of the period to consider

max
numeric. The end of the period to consider

type
character. The time units of min and max

Details

The limits of the period to consider may correspond to any of the components of the list of class POSIXlt (hour, day, month, etc.; see help(POSIXlt)), or to dates stored in objects of class POSIXct (see examples).

Value

an object of class ltraj.

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

ltraj for further information about objects of class ltraj, POSIXlt for further information about objects of class POSIXlt

Examples

data(puechcirc)
plot(puechcirc, perani = FALSE)

## Gets all the relocations collected
## between midnight and 3H AM 
toto <- gdltraj(puechcirc, min = 0, max = 3, type="hour")
plot(toto, perani = FALSE)

## Gets all relocations collected between the 15th
## and the 25th august 1993
lim <- as.POSIXct(strptime(c("15/08/1993", "25/08/1993"),
                  "%d/%m/%Y"))
tutu <- gdltraj(puechcirc, min = lim[1],
                max = lim[2], type="POSIXct")
plot(tutu, perani = FALSE)

[Package adehabitat version 1.8.18 Index]