getXYcoords {adehabitat}	R Documentation

Computes the X and Y Coordinates of the Pixels of a Raster Map

Description

getXYcoords computes the geographical coordinates of the rows and columns of pixels of a raster map.

Usage

getXYcoords(w)

Arguments

w
an object of class asc, kasc, sahrlocs, or mapattr.

Value

Returns a list with two components:

x
the x coordinates of the columns of pixels of the map

y
the y coordinates of the rows of pixels of the map

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

asc, kasc, as.sahrlocs, storemapattr for additionnal informations on objects of class area, kasc, sahrlocs, mapattr respectively

Examples

data(puechabon)
(elev <- getkasc(puechabon$kasc, "Elevation"))
(coords <- getXYcoords(elev))
nrow(elev) == length(coords$x)
ncol(elev) == length(coords$y)

[Package adehabitat version 1.8.18 Index]