hbrown {adehabitat}	R Documentation

Estimates the value of h for a Brownian motion

Description

hbrown estimates the scaling factor h (used in the Brownian motion, see help(simm.brown)) from a trajectory.

Usage

hbrown(x)

Arguments

x
an object of class ltraj

Value

a vector with one estimate per burst of the object of class ltraj.

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

simm.brown

Examples

toto <- simm.brown(1:200, h=4)
hbrown(toto)

toto <- simm.brown(1:200, h=20)
hbrown(toto)

[Package adehabitat version 1.8.18 Index]