hist.kasc {adehabitat}	R Documentation

Histograms of Mapped Variables

Description

hist.kasc performs histograms of the variables mapped in objects of class kasc.

Usage

## S3 method for class 'kasc'
hist(x, type = c("h", "l", "b"), adjust = 1, col,
          border, lwd = 1, ...)

Arguments

x
a raster map of class kasc

type
what type of plot should be drawn. Possible types are:
* "h" for histograms,
* "l" for kernel density estimates (see ?density).
* "b" for both histograms and kernel density estimates (see ?density).
By default, type = "h" is used. If type = "l" is used, the position of the mean of each distribution is indicated by dotted lines

adjust
if type = "l", a parameter used to control the bandwidth of the density estimate (see ?density)

col
color for the histogram

border
color for the border of the histogram

lwd
if type = "l", line width for the density estimate

...
further arguments passed to or from other methods

Author(s)

Mathieu Basille basille@ase-research.org

See Also

kasc

Examples

## Example with factors and numeric variables
data(puechabon)
hist(puechabon$kasc, type = "h")

## Aspect is a factor, then it's not possible to use 
## kernel density estimates for it :
hist(puechabon$kasc)

## Removing the factor Aspect, and smoothing gives :
hist(puechabon$kasc[-2], type = "l")

[Package adehabitat version 1.8.18 Index]