hist.ltraj {adehabitat}	R Documentation

Histogram of the Descriptive Parameters of a Trajectory

Description

This function draws an histogram of any tranformation of the descriptive parameters of a trajectory in objects of class ltraj.

Usage

## S3 method for class 'ltraj'
hist(x, which = "dx/sqrt(dt)", ...)

Arguments

x
an object of class ltraj

which
a character string giving any syntactically correct R expression implying the descriptive elements in x.

...
parameters to be passed to the generic function hist.

Value

a list of objects of class "histogram"

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

hist, ltraj for additional information on the descriptive parameters of the trajectory, qqnorm.ltraj for examination of distribution.

Examples

## Simulation of a Brownian Motion
a <- simm.brown(c(1:300, seq(301,6000,by=20)))
plot(a, addpoints = FALSE)

## dx/sqrt(dt) and dy/sqrt(dt) are normally distributed (see
## ?qqchi) 
hist(a, "dx/sqrt(dt)", freq = FALSE)
lines(tutu <- seq(-5,5, length=50), dnorm(tutu), col="red")

hist(a, "dy/sqrt(dt)", freq = FALSE)
lines(tutu, dnorm(tutu), col="red")

## Look at the distribution of distances between
## successive relocations
hist(a, "dist/sqrt(dt)", freq = FALSE)
lines(tutu <- seq(0,5, length=50), dchi(tutu), col="red")

[Package adehabitat version 1.8.18 Index]