hr.rast {adehabitat}	R Documentation

Rasterisation of Objects of Class 'area'

Description

Converts an object of class area (used in many functions of the package ade4) to an object of class kasc (rasterisation).

Usage

hr.rast(mcp, w, border=c("include", "exclude"))

Arguments

mcp
an object of class area

w
a raster map of class kasc or of class asc

border
a character string indicating what happens when the center of the pixel is located exactly on the limit of the polygon ("include" indicates that the pixel is considered to be inside the polygon).

Value

Returns an object of class kasc.

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

kasc for additional information on objects of class kasc, area for further information on the class area

Examples

data(puechabon)
kasc <- puechabon$kasc
locs <-  puechabon$locs

## Computes the home range of the animals
cp <- mcp(locs[,4:5], locs[,1])
area.plot(cp)

## Converts the home range to raster
cprast <- hr.rast(cp, kasc)
image(cprast)

[Package adehabitat version 1.8.18 Index]