hseal {adehabitat}	R Documentation

Argos Monitoring of Hooded Seal

Description

This data set contains the trajectory of one hooded seal.

Usage

data(hseal)

Format

The dataset hseal is an object of class ltraj. The coordinates are stored in meters (UTM - zone 21).

Source

Jonsen, I.D., Flemming, J.M. and Myers, R.A. (2005). Robust state-space modeling of animal movement data. Ecology, 86, 2874–2880.

Examples

data(hseal)

plot(hseal)

[Package adehabitat version 1.8.18 Index]