ibex {adehabitat}	R Documentation

GPS Monitoring of Four Ibex in the Belledonne Mountain

Description

This dataset is an object of class "ltraj" (regular trajectory, relocations every 4 hours) containing the GPS relocations of four ibex during 15 days in the Belledonne mountain (French Alps).

Usage

data(ibex)

Source

Office national de la chasse et de la faune sauvage, CNERA Faune de Montagne, 95 rue Pierre Flourens, 34000 Montpellier, France.

Examples

data(ibex)
plot(ibex)

[Package adehabitat version 1.8.18 Index]