image.sahrlocs {adehabitat}	R Documentation

Graphical Display of the Habitat Composition of the Home Ranges of Animals Monitored Using Radio-Tracking

Description

image.sahrlocs allows a gray-level display of the composition of home ranges (different colors are used for factors). For a given variable, the minimum gray level (default is "white") and the maximum gray level (default is "black") represents respectively the minimum and the maximum of the variable ***on the study area***.

Usage

## S3 method for class 'sahrlocs'
image(x, ani = names(x$hr), var = names(x$sa),
               mar = c(0, 0, 0, 0), axes = FALSE, dfidxy = NULL, colpts =
               "black", pch = 21, bg = "white", inv = FALSE, cexpts = 0.6, 
               csub = 2, possub = c("bottomleft", "bottomright", "topleft",
               "topright"), ...)

Arguments

x
an object of class sahrlocs

ani
a character vector giving the names of the variables of the "hr" component (the animals) for which a display is wanted

var
a character vector giving the names of the variables of the "sa" component (the habitat variables) for which a display is wanted

mar
the graphical parameter mar (see par)

axes
logical. Whether the axes should be plotted

csub
the character size for the legend, used with par("cex")*csub

possub
a character string indicating the sub-title position ("topleft", "topright", "bottomleft", "bottomright")

dfidxy
an optional data frame with three columns giving the identity and the coordinates of the relocations of each animal. (if not NULL, the relocations of each animal are plotted in its home range, see Examples)

colpts
if dfidxy is not NULL, the color of the points to be used for the plot of the relocations

pch
if dfidxy is not NULL, the size of the points to be used for the plot of the relocations (see par)

bg
if dfidxy is not NULL, the background color to be used for the plot of the relocations (see par)

inv
by default, lower values of the mapped variables are brighter. If FALSE, the lower values are darker

cexpts
if dfidxy is not NULL, the size of the points

...
additionnal parameters to be passed to the generic function image

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

as.sahrlocs for additionnal information on objects of class sahrlocs

Examples

data(puechabon)
(sahr <- puechabon$sahr)

## Displays all the variables for a given animal
image(sahr, ani = "Chou")

## Displays all the animals for a given variable
image(sahr, var = "Elevation")

## Load and displays the relocations of the animals
locs <-  puechabon$locs[,c(1,4:5)]
image(sahr, var = "Elevation", dfidxy = locs, pch = 21)

[Package adehabitat version 1.8.18 Index]