join.asc {adehabitat}	R Documentation

Finds the Value of Mapped Variables at some Specified Locations (Spatial Join)

Description

join.asc finds the value of a mapped variable at some specified locations.
join.kasc is the same function as join.asc, with several maps.

Usage

 join.asc(pts, x) 
 join.kasc(pts, w) 

Arguments

x
an object of class asc

w
an object of class kasc

pts
a data frame with two columns containing the x and y coordinates of the points to be placed on the map

Value

join.asc returns a vector with length equals to the number of points in pts.
join.kasc returns a data frame with a number of columns equals to the number of variables in the object of class kasc, and with each row corresponding to the rows of pts.

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

kasc and asc

Examples

data(puechabon)
x <- puechabon$kasc

## for each relocation, finds the values of the variables in x
toto <- join.kasc(pts = cbind(puechabon$locs$X, puechabon$locs$Y), x)
toto[1:4,]

[Package adehabitat version 1.8.18 Index]