kerneloverlap {adehabitat}	R Documentation

Spatial Interaction between Animals Monitored Using Radio-Tracking

Description

These functions implements all the indices of kernel home-range overlap reviewed by Fieberg and Kochanny (2005). kerneloverlap computes these indices from a set of relocations, whereas kerneloverlaphr computes these indices from an object containing the utilization distributions of the animals.

Usage

kerneloverlap(xy, id = NULL,
              method = c("HR", "PHR", "VI", "BA", "UDOI", "HD"),
              lev = 95, conditional = FALSE, ...)

kerneloverlaphr(x, method = c("HR", "PHR", "VI", "BA", "UDOI", "HD"),
                lev=95, conditional=FALSE)

Arguments

xy
a data frame with two columns (x and y coordinates of the animal relocations)

id
a factor giving the animals identity associated to xy

x
an object of class khrud containing the utilization distributions of several animals, and returned by the function kernelUD.

method
the desired method for the estimation of overlap (see details)

lev
the level of the home range estimation

conditional
logical. If TRUE, the function sets to 0 the pixels of the grid over which the UD is estimated, outside the home range of the animal estimated at a level of probability equal to lev. Note that this argument has no effect when meth="HR".

...
additional arguments to be passed to the function kernelUD for the kernel estimation of the utilization distribution.

Details

Fieberg and Kochanny (2005) made an extensive review of the indices of overlap between utilization distributions (UD) of two animals. The function kerneloverlap implements these indices. The argument method allows to choose an index

The choice method="HR" computes the proportion of the home range of one animal covered by the home range of another one, i.e.:

HR[i,j] = A[i,j] / A[i]

, where A[i,j] is the area of the intersection between the two home ranges and A[i] is the area of the home range of the animal i.

The choice method="PHR" computes the volume under the UD of the animal j that is inside the home range of the animal i (i.e., the probability to find the animal j in the home range of i). That is:

PHR[i,j] = integral_over_x integral_over_y A[i] UD[j](x,y) dxdy

where UD[j](x,y) is the value of the utilization distribution of the animal j at the point x,y.

The choice method="VI" computes the volume of the intersection between the two UD, i.e.:

VI = integral_over_x integral_over_y min(UD[i](x,y),UD[j](x,y)) dxdy

Other choices rely on the computation of the joint distribution of the two animals under the hypothesis of independence UD[i](x,y) * UD[j](x,y).

The choice method="BA" computes the Bhattacharyya's affinity

BA = integral_over_x integral_over_y sqrt(UD[i](x,y)) * sqrt(UD[j](x,y))

The choice method="UDOI" computes a measure similar to the Hurlbert index of niche overlap:

UDOI = A[i,j] integral_over_x integral_over_y UD[i](x,y) * UD[j](x,y)

The choice method="HD" computes the Hellinger's distance:

HD = integral_over_x integral_over_y ((sqrt(UD[i](x,y) - sqrt(UD[j](x,y)))^2 dxdy)^(1/2))

Value

A matrix giving the value of indices of overlap for all pairs of animals.

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr, based on a work of John Fieberg

References

Fieberg, J. and Kochanny, C.O. (2005) Quantifying home-range overlap: the importance of the utilization distribution. Journal of Wildlife Management, 69, 1346–1359.

See Also

kernelUD for additional information on kernel estimation of home ranges

Examples

## Not run: 
data(puechabon)

kerneloverlap(puechabon$locs[,c("X","Y")], puechabon$locs$Name,
              grid=200, meth="VI", conditional=TRUE)

## Identical to
kud <- kernelUD(puechabon$locs[,c("X","Y")], puechabon$locs$Name,
                grid=200, same4all=TRUE)
kerneloverlaphr(kud, meth="VI", conditional=TRUE)

## Other indices:

kerneloverlap(puechabon$locs[,c("X","Y")], puechabon$locs$Name,
              grid=200, meth="HR")

kerneloverlap(puechabon$locs[,c("X","Y")], puechabon$locs$Name,
              grid=200, meth="PHR")

kerneloverlap(puechabon$locs[,c("X","Y")], puechabon$locs$Name,
              grid=200, meth="BA")

kerneloverlap(puechabon$locs[,c("X","Y")], puechabon$locs$Name,
              grid=200, meth="UDOI")

kerneloverlap(puechabon$locs[,c("X","Y")], puechabon$locs$Name,
              grid=200, meth="HD")

## End(Not run)

[Package adehabitat version 1.8.18 Index]