kver {adehabitat}	R Documentation

Handling of Objects of Class kver

Description

Objects of class kver are created by getverticeshr and getverticesclusthr: they contain the vertices of the home range estimated using either nearest neighbour clustering or kernel estimation. An object of class kver is list of data frames of class area.

plot.kver is used to display the home-range contours defined in an object of clss kver.

kver.rast is used to rasterize the home ranges.

kver2shapefile is used to convert the home ranges into a shapefile object (for exportation to a GIS).

Usage

kver.rast(kv, asc)
kver2shapefile(kv, which = names(kv))
## S3 method for class 'kver'
plot(x, which = names(x), colpol = rainbow(length(which)),
          colborder = rep("black", length(which)), lwd = 2,
          add = FALSE, ...)

Arguments

kv,x
an object of class kver

asc
a matrix of class asc

which
a vector of character indicating the polygons to be plotted

colpol
a vector of the color for filling the polygon. The default, NA, is to leave polygons unfilled

colborder
a vector of the color to draw the border (black, by default). Use border = NA to omit borders

lwd
the border width, a positive number

add
logical. If TRUE, the polygons are added to a previous plot

...
additional arguments to be passed to the function plot.area

Value

kver.rast returns an object of class asc.
kver2shapefile returns a shapefile object (can be exported using the function write.shapefile of the package shapefiles.

Warning

kver2shapefile requires the packages shapefiles.

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

getverticesclusthr and getverticeshr for functions creating this class of objects. area for more information on objects of class area. asc for more information on objects of class asc. write.shapefile for exportation of shape files.

Examples

data(puechabon)
lo<-puechabon$locs[,c("X","Y")]

## Home Range Estimation
res <- clusthr(lo, puechabon$locs$Name)

## gets the vertices
vec <- getverticesclusthr(res)
plot(vec)
image(kver.rast(vec, getkasc(puechabon$kasc,1)))

[Package adehabitat version 1.8.18 Index]