lynxjura {adehabitat}	R Documentation

Monitoring of Lynx

Description

This data set stores the results of the monitoring of lynx in the French Jura between 1980 and 1999. These data have been collected by the Lynx Network of the french wildlife management office (Office national de la chasse et de la faune sauvage).

Usage

data(lynxjura)

Format

The list lynxjura has two components:

mapan object of class kasc (see help(kasc)) that describes several variables on the study area: forets is the density of forests, hydro is the density of rivers, routes is the density of roads and artif is the distance from urbanized areas.

locsa data frame containing the locations of presence indices of the lynx. X and Y are the x and y coordinates, Date is the date of the collection of the indice and Type represents the type of data (C: alive lynx captured, D: attacks on livestock, E: prints or tracks, F: feces, J: hairs, L: corpse of lynx, O: sightings and P: attacks on wild prey).

Source

Vandel, J.M. (2001) Repartition du Lynx (Lynx lynx) en France (Massif Alpin, Jurassien et Vosgien). Methodologie d'etude et statut actuel. Ecole Pratique des Haute Etudes de Montpellier II: Dissertation.

[Package adehabitat version 1.8.18 Index]