managNAkasc {adehabitat}	R Documentation

"Cleaning" Objects of Class 'kasc'

Description

An object of class kasc stores several maps in a data frame (one column per variable, and one row per pixel of the raster map). However, the features mapped are rarely rectangle-shaped, whereas the map are inevitably rectangles. Therefore, a lot of pixels of the maps do not contain data. The pixels of the map that do not contain data are NA in this data frame.
It often occurs that several variables are not mapped for exactly the same area (that is, some pixels are NA for some variables, and not for others). managNAkasc will set to NA all pixels that are not mapped for all variables.

Usage

managNAkasc(x)

Arguments

x
an object of class kasc

Value

Returns an object of class kasc.

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

kasc for additional information on objects of class kasc.

[Package adehabitat version 1.8.18 Index]