mcp {adehabitat}	R Documentation

Estimation of the Home Range Using the Minimum Convex Polygon Estimator

Description

mcp computes the home range of several animals using the Minimum Convex Polygon estimator.
mcp.area is used for home-range size estimation.
plot.hrsize is used to display the home-range size estimated at various levels.

Usage

mcp(xy, id, percent = 95)
mcp.area(xy, id, percent = seq(20,100, by = 5),
         unin = c("m", "km"),
         unout = c("ha", "km2", "m2"), plotit = TRUE)
## S3 method for class 'hrsize'
plot(x, ...)

Arguments

xy
a data frame with two columns containing the coordinates of the relocations of the monitored animals

id
a factor giving the identity of the animal for each relocation

percent
100 minus the proportion of outliers to be excluded from the computation

unin
the units of the relocations coordinates. Either "m" (default) for meters or "km" for kilometers

unout
the units of the output areas. Either "m2" for square meters, "km2" for square kilometers or "ha" for hectares (default)

plotit
logical. Whether the plot should be drawn.

x
an objet of class hrsize returned by the function mcp.area, or kernel.area (see kernelUD())

...
additionnal arguments to be passed to the function plot

Details

This function computes the Minimum Convex Polygon estimation after the removal of (100 minus percent) percent of the relocations the farthest away from the barycenter of the home range (computed by the arithmetic mean of the coordinates of the relocations for each animal).

Value

mcp returns an object of class area, with one polygon per level of the factor ID.
mcp.area returns a data frame of class hrsize, with one column per animal and one row per level of estimation of the home range.

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

References

Mohr, C.O. (1947) Table of equivalent populations of north american small mammals. The American Midland Naturalist, 37, 223-249.

See Also

chull, plot.area to have a graphical display of the home ranges, area for additionnal information on the class area, and area2dxf for further exportation toward a GIS. s.chull for another way to display MCP.

Examples

data(puechabon)
locs <-  puechabon$locs

cp <- mcp(locs[,4:5], locs[,1])

## Plot the home ranges
opar <- par(mar = c(0,0,0,0))
area.plot(cp)

## ... And the relocations
points(locs[,4:5], pch = 16, col = as.numeric(locs[,1]))
par(opar)

## Computation of the home-range size:
cuicui1 <- mcp.area(locs[,4:5], locs[,1])
plot(cuicui1)

[Package adehabitat version 1.8.18 Index]