mcp.rast {adehabitat}	R Documentation

Converts a Polygon to Raster

Description

mcp.rast converts a polygon map to a raster map of class asc.

Usage

  mcp.rast(poly, w, border=c("include", "exclude"))

Arguments

poly
a data frame with 2 columns giving the coordinates of a polygon object

w
an object of class kasc, or of class asc

border
a character string indicating what happens when the center of the pixel is located exactly on the limit of the polygon ("include" indicates that the pixel is considered to be inside the polygon).

Details

The raster map is needed to pass the format for the output raster object to the function.

Value

Returns an object of class asc.

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

hr.rast

Examples

data(puechabon)
toto <- puechabon$kasc
loc <- puechabon$locs

## gets the coordinates of the relocations for the wild boar #1
wb1 <- loc[loc$Name == "Chou",]
wb1 <- cbind(wb1$X, wb1$Y)
nbpol <- chull(wb1)
xycoord <- wb1[nbpol,]

## rasterization of wb1
tutu <- mcp.rast(xycoord, toto)
image(tutu)

polygon(xycoord, lwd = 2)

[Package adehabitat version 1.8.18 Index]