mindistkeep {adehabitat}	R Documentation

Detecting Absence of Movement in an Object of Class 'ltraj'

Description

Objects of class ltraj are often created with data collected using some form of telemetry (radio-tracking, G.P.S., etc.). However, the relocations of the monitored animals are always somewhat imprecise. The function mindistkeep considers that when the distance between two successive relocations is lower than a given threshold distance, the animal actually does not move.

Usage

mindistkeep(x, threshold)

Arguments

x
An object of class ltraj

threshold
The minimum distance under which is is considered that the animal does not move

Value

An object of class ltraj

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

ltraj

Examples

data(puechcirc)
plot(puechcirc)

i <- mindistkeep(puechcirc, 10)
plot(i)

[Package adehabitat version 1.8.18 Index]