morphology {adehabitat}	R Documentation

Morphology: Erosion or Dilatation of Features on a Raster Map

Description

morphology performs morphological operations on images of class asc.

Usage

morphology(x, operation = c("erode", "dilate"), nt = 5)

Arguments

x
a matrix of class asc

operation
a character string indicating the operation to be processed: either "erode" or "dilate"

nt
the number of times that the operation should be processed

Value

Returns a matrix of class asc, containing 1 when the pixel belong to one feature of the image and NA otherwise (see examples).

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

asc for further information on objects of class asc.

Examples

data(puechabon)
a <- getkasc(puechabon$kasc,"Elevation")

## dilatation
toto1 <- morphology(a, operation = "dilate", nt = 1)
toto2 <- morphology(a, operation = "dilate", nt = 2)
toto3 <- morphology(a, operation = "dilate", nt = 3)
toto5 <- morphology(a, operation = "dilate", nt = 5)
colo <- grey((1:5)/6)
image(toto5, col = colo[1])
image(toto3, add = TRUE, col = colo[2])
image(toto2, add = TRUE, col = colo[3])
image(toto1, add = TRUE, col = colo[4])
image(a, add = TRUE)

## erosion
colo <- grey((1:20)/21)
image(a, col = 1)
for (i in 1:19) {
  toto <- morphology(a, operation = "erode", nt = i)
  image(toto, add = TRUE, col = colo[i])
}

[Package adehabitat version 1.8.18 Index]