mouflon {adehabitat}	R Documentation

GPS Monitoring of One Mouflon in the Caroux Mountain

Description

This dataset is an object of class "ltraj" (regular trajectory, relocations every 20 minutes) containing the GPS relocations of one mouflon during two week-ends in the Caroux mountain (South of France).

Usage

data(mouflon)

Source

Office national de la chasse et de la faune sauvage, CNERA Faune de Montagne, 95 rue Pierre Flourens, 34000 Montpellier, France.

Examples

data(mouflon)
plot(mouflon)

[Package adehabitat version 1.8.18 Index]