perarea {adehabitat}	R Documentation

Compute Areas and Perimeters of Objects of Class "area"

Description

perarea computes the perimeters of polygons in objects of class area.
ararea computes the areas of polygons in objects of class area.

Usage

perarea(x)
ararea(x)

Arguments

x
object of class area

Value

a vector.

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

area, plot.area

Examples

data(puechabon)
locs <-  puechabon$locs

cp <- mcp(locs[,4:5], locs[,1])

perarea(cp)

ararea(cp)

[Package adehabitat version 1.8.18 Index]