pheasant {adehabitat}	R Documentation

Radio-Tracking of Pheasants

Description

This data set describes the use and availability of 5 habitat types for 13 pheasants monitored using radio-tracking.

Usage

data(pheasant)

Format

This list has three components:

studyarea

a data frame giving the proportion of each habitat type (columns) on the study area. These habitat types are Scrub, Broadleaf, Coniferous, Grassland and Crop. These proportions are repeated by rows, for all animals (rows)

mcp

a data frame giving the proportion of each habitat type (columns) in the home range of each animal (rows)

locs

a data frame giving the proportion of relocations of each animal (rows) reported in 3 of the 5 habitat types (columns). Indeed, Coniferous and Crops were not used by most of the animals.

Source

Aebischer, N. J., Robertson, P. A. and Kenward, R. E. (1993) Compositional analysis of habitat use from animal radiotracking data. Ecology, 74, 1313–1325.

[Package adehabitat version 1.8.18 Index]