plot.ltraj {adehabitat}	R Documentation

Graphical Display of an Object of Class "ltraj"

Description

plot.ltraj allows various graphical displays of the trajectories.

Usage

## S3 method for class 'ltraj'
plot(x, id = unique(unlist(lapply(x, attr, which = "id"))),
           burst = unlist(lapply(x, attr, which = "burst")), asc = NULL,
           area = NULL, xlim = NULL, ylim = NULL, colasc =
           gray((240:1)/256), colpol = "green",  addpoints = TRUE,
           addlines = TRUE, perani = TRUE, final = TRUE, ...)

Arguments

x
an object of class ltraj

id
a character vector containing the identity of the individuals of interest

burst
a character vector containing the burst levels of interest

asc
an object of class asc

area
an object of class area

xlim
the ranges to be encompassed by the x axis

ylim
the ranges to be encompassed by the y axis

colasc
a character vector giving the colors of the map of class asc

colpol
a character vector giving the colors of the polygon contour map, when area is not NULL

addpoints
logical. If TRUE, points corresponding to each relocation are drawn

addlines
logical. If TRUE, points corresponding to each relocation are drawn

perani
logical. If TRUE, one plot is drawn for each value of id, and the several bursts are superposed on the same plot for a given animal. If FALSE, one plot is drawn for each value of burst

final
logical. If TRUE, the initial and final relocations of each burst are indicated in blue and red, respectively

...
arguments to be passed to the generic function plot

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

For further information on the class ltraj, ltraj.

Examples

data(puechcirc)

plot(puechcirc)
plot(puechcirc, perani = FALSE)
plot(puechcirc, id = "JE93", perani = FALSE)

data(puechabon)
ele <- getkasc(puechabon$kasc, "Elevation")
plot(puechcirc, perani = FALSE, asc = ele)

cont <- getcontour(ele)
plot(puechcirc, area = cont)

[Package adehabitat version 1.8.18 Index]