plot.sahrlocs {adehabitat}	R Documentation

Exploratory Analysis of Habitat Selection

Description

plot.sahrlocs applies the function widesII or widesIII for each variable in the object of class sahrlocs, and the results are stored in a list. Then, the function plot.wi is applied to each component of the list. This allows to investigate habitat selection by animals at several scales for design II and III data.

Usage

## S3 method for class 'sahrlocs'
plot(x, ani = names(x$hr), var = names(x$sa),
              type = c("hr.in.sa", "locs.in.hr", "locs.in.sa"),
              ncla = 4, ylog = FALSE, caxis = 0.7, clab = 0.7,
              errbar = c("SE", "CI"), alpha = 0.05, draw = TRUE, ...)

Arguments

x
an object of class sahrlocs

ani
a character vector. This vector contains the names of the animals in x for which habitat selection should be displayed. At least two animals are required

var
a character vector. This vector contains the names of the variables in x for which habitat selection should be displayed

type
a character string. Type of habitat selection that should be investigated. If "hr.in.sa", the selection of the home-range within the study area is displayed. If "locs.in.sa", the selection of the relocations within the study area is displayed. If "locs.in.hr", the selection of the relocations within the home range is displayed

ncla
numeric variables are converted to factors. This parameter controls the number of classes of these factors

ylog
logical. If TRUE, the selection ratios are plotted on a log scale

caxis
character size on axes to be passed to par("cex.axis")

clab
character size of axes labels to be passed to par("cex.lab")

errbar
a character string. Type of error bars: either "CI" for confidence intervals or "SE" for standard errors

alpha
the alpha-level for the tests

draw
logical. If FALSE, no plot is drawn

...
further arguments to be passed to the function plot.wi

Value

The function returns a list of objects of class wiII or wiIII (one component per animal).

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

References

Manly B.F.J., McDonald L.L., Thomas, D.L., McDonald, T.L. & Erickson, W.P. (2003) Resource selection by animals - Statistical design and Analysis for field studies. Second edition. London: Kluwer academic publishers.

See Also

widesII and widesIII for further information on objects of class wiII and wiIII, as.sahrlocs for further information on objects of class sahrlocs.

Examples

data(puechabon)
sahr <- puechabon$sahr
toto <- plot(sahr)

### Note that the wild boars named Brock and Calou
### have only one herbaceous cover class available
### (the second one), and they use it exclusively.
### So they have identical
### selection ratios (that's why the curve of Brock
### does not appear: it is hidden behind the curve
### of the boar named Calou).

toto
toto$Elevation

[Package adehabitat version 1.8.18 Index]