plotltr {adehabitat}	R Documentation

Changes in Traject Parameters Over Time

Description

This function allows a graphical examination of the changes in descriptive parameters in objects of class ltraj

Usage

plotltr(x, which = "dist", ...)

Arguments

x
An object of class ltraj

which
a character string giving any syntactically correct R expression implying the descriptive elements in x.

...
additional parameters to be passed to the generic function plot

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

ltraj for additional information about objects of class ltraj, and sliwinltr for a sliding window smoothing

Examples

data(puechcirc)

plotltr(puechcirc, "cos(rel.angle)")
plotltr(puechcirc, "dist")
plotltr(puechcirc, "dx")

[Package adehabitat version 1.8.18 Index]