puech {adehabitat}	R Documentation

Radio-Tracking Data of Wild Boar (2)

Description

This data set stores the results of the monitoring of 6 wild boar at Puechabon (Mediterranean habitat, South of France). These data have been collected by Daniel Maillard (Office national de la chasse et de la faune sauvage).

Usage

data(puech)

Details

The list puech has two components:
puech$relocations is a data frame containing the relocations of the wild boar resting sites in summer. It contains the coordinates of the relocations and the name of the corresponding wild boar.
puech$map is a list of objects of class asc (see help(asc)) that describe nine environmental variables on the study area (the elevation, the tree cover, the shrub cover, the distance to recreational trails, the distance to crops, the distance to water points, the grass cover, the slope and the sunshine).
Note that both the maps and the relocations have been slightly "damaged" to preserve the copyright on the data.

References

Maillard, D. (1996). Occupation et utilisation de la garrigue et du vignoble mediterraneens par le Sanglier. Universite d'Aix-Marseille III: PhD thesis.

[Package adehabitat version 1.8.18 Index]