puechabon {adehabitat}	R Documentation

Radio-Tracking Data of Wild Boar

Description

This data set stores the results of the monitoring of 4 wild boars at Puechabon (Mediterranean habitat, South of France). These data have been collected by Daniel Maillard (Office national de la chasse et de la faune sauvage).

Usage

data(puechabon)

Details

The list puechabon has three components:
puechabon$kasc is an object of class kasc (see help(kasc)) that describes several variables on the study area.
locs is a data frame of the relocations of the wild boar resting sites in summer. Information on wild boars is provided by factors Name, Sex, Age.
sahr is the associated object of class sahrlocs. The home ranges were estimated with buffers including all pixels within 500 m of a boar relocation.

References

Maillard, D. (1996). Occupation et utilisation de la garrigue et du vignoble mediterraneens par le Sanglier. Universite d'Aix-Marseille III: PhD thesis.

[Package adehabitat version 1.8.18 Index]