puechcirc {adehabitat}	R Documentation

Movements of wild boars tracked at Puechabon

Description

This data set is an object of class ltraj, giving the results of the monitoring of 2 wild boars by radio-tracking at Puechabon (Mediterranean habitat, South of France). These data have been collected by Daniel Maillard (Office national de la chasse et de la faune sauvage), and correspond to the activity period of the wild boar (during the night, when the animals forage. The data set puechabon describes the resting sites).

Usage

data(puechcirc)

Format

This object has, in total, 204 relocations distributed among two animals and three bursts of relocations (CH930803, CH930824, CH930827, and JE930827).

Source

Maillard, D. (1996). Occupation et utilisation de la garrigue et du vignoble mediterraneens par le Sanglier. Universite d'Aix-Marseille III: PhD thesis.

[Package adehabitat version 1.8.18 Index]