puechdesIII {adehabitat}	R Documentation

Habitat Selection by the Wild Boar at Puechabon

Description

This data set contains two data frames describing the use and the availability of 3 elevation classes for 6 wild boars (Sus scrofa L.) monitored using radio-tracking at Puechabon (South of France). These data have been collected by Daniel Maillard (Office national de la chasse et de la faune sauvage).

Usage

data(puechdesIII)

Details

The list puechdesIII has two components:
The data frame used describes the number of telemetry relocations for each of the 6 animals in each of the three elevation classes.
The data frame available describes a sample of random points placed in the areas available to these wild boars (a buffer area of 200 m around the relocations).

Source

Maillard, D. (1996) Occupation et utilisation de la garrigue et du vignoble mediterraneens par le Sanglier. Universite d'Aix-Marseille III: PhD thesis.

[Package adehabitat version 1.8.18 Index]