qqchi {adehabitat}	R Documentation

Quantile-Quantile Plots for Trajectories of Class 'ltraj'

Description

The functions allow the examination of the distribution of trajectories descriptors (see Details).

Usage

## Chi distribution of the increment length / sqrt(dt)
qqchi(y, ...)

## Default S3 method:
qqchi(y, df = 2, ylim, main = "Chi Q-Q Plot",
              xlab = "Theoretical Quantiles", ylab = "Sample Quantiles",
              plot.it = TRUE, datax = FALSE, ...) 

## S3 method for class 'ltraj'
qqchi(y, xlab = "Theoretical Quantiles", 
            ylab = "Sample Quantiles (Distances)", ...)

## Normal Distribution of dx/sqrt(dt) or dy/sqrt(dt)
## S3 method for class 'ltraj'
qqnorm(y, which=c("dx","dy"), ...)

Arguments

y
a vector containing the data sample for qqchi.default. an object of class ltraj for other functions.

df
the number of degrees of freedom of the Chi distribution (default to 2).

xlab, ylab, main
plot labels.

plot.it
logical. Should the result be plotted?

datax
logical. Should data values be on the x-axis?

which
a character string indicating the component (dx or dy) to be examined.

ylim, ...
graphical parameters.

Details

Among the numerous statistics that can be used to describe the movements of an animal, the length of the increment between two successive relocations is very common. This increment can be described by a vector i = c(dx, dy). Under the hypothesis of a Brownian motion, dx and dy should be normally distributed with mean = 0 and variance = dt (where dt is the time interval between the two relocations). Therefore, dx/sqrt(dt) and dy/sqrt(dt) should be normally distributed with mean = 0 and variance = 1. The function qqnorm.ltraj performs a quantile-quantile plot of dx/sqrt(dt) or dy/sqrt(dt) vs. a normal distribution to verify wether the Brownian motion assumption is correct.

Furthermore, the quantity (dx^2 + dy^2)/dt should be distributed according to a Chi-squared distribution with two degrees of freedom. Thus, the quantity distance / sqrt(dt) should be distributed according to a Chi distribution with two degrees of freedom (where distance is the distance between the two relocations). The function qqchi.ltraj performs quantile-quantile plot of distance/sqrt(dt) vs. a Chi distribution to verify wether the Brownian motion assumption is correct.

Value

for functions dealing with objects of class ltraj, a list with components being themselves lists, with components:

x
The x coordinates of the points that were/would be plotted

y
The original y vector, i.e., the corresponding y coordinates including 'NA's.

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

chi, qqplot, ltraj.

Examples

## Example with an Arithmetic Brownian Process
toto <- simm.mba(1:500, sig = diag(c(5, 5)))
qqnorm(toto, "dx")
qqnorm(toto, "dy")
qqchi(toto)

## Example of wild boar
data(puechcirc)
qqnorm(puechcirc, "dx")
qqnorm(puechcirc, "dy")
qqchi(puechcirc)

[Package adehabitat version 1.8.18 Index]