randtest.enfa {adehabitat}	R Documentation

Randomisation Test for the Ecological Niche Factor Analysis

Description

randtest.enfa performs a randomisation test for the Ecological Niche Factor analysis (ENFA).

Usage

## S3 method for class 'enfa'
randtest(xtest, nrepet = 999, ...)

Arguments

xtest
an object of class enfa

nrepet
the number of iterations for the randomisation test

...
further arguments to be passed to the generic function randtest

Details

This test is carried out by simulating a random distribution of the species occurrences in the pixels of a map.

At each step of the randomisation procedure, the test randomly allocates the nk occurrences (where nk is the sum of the occurrence vector pr of the object of class enfa) in the Ik pixels of the focus area (where Ik is the length of this occurrence vector).

At each step of the procedure, the first eigenvalue of the ENFA performed on the randomised data set is recomputed. This value provides a criterion to test the pertinence of the ENFA analysis.

Value

returns a list of class randtest

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

References

Manly, B.F.J. (1997) Randomization, Bootstrap and Monte Carlo Methods in Biology. London: Chapman & Hall.

Hirzel, A.H., Hausser, J., Chessel, D. and Perrin, N. (2002) Ecological-niche factor analysis: How to compute habitat suitability maps without absence data? Ecology, 83, 2027–2036.

See Also

enfa

Examples

## Not run: 
data(chamois)
locs <- chamois$locs
map <- chamois$map
map$Vegetation <- NULL
da <- data2enfa(map, locs)
en <- enfa(dudi.pca(da$tab, scannf=FALSE), da$pr, scannf = FALSE)
(tutu <- randtest(en, nrepet = 100))
plot(tutu)

## End(Not run)

[Package adehabitat version 1.8.18 Index]