rupicabau {adehabitat}	R Documentation

GPS Monitoring of One Chamois in the Bauges Mountains

Description

This dataset is an object of class "ltraj" (regular trajectory, relocations every 20 minutes) containing the GPS relocations of two chamois during one day in the Bauges mountain (French Alps).

Usage

data(rupicabau)

Source

Office national de la chasse et de la faune sauvage, CNERA Faune de Montagne, 95 rue Pierre Flourens, 34000 Montpellier, France.

Examples

data(rupicabau)
plot(rupicabau)

[Package adehabitat version 1.8.18 Index]