scatterniche {adehabitat}	R Documentation

Display the Niche in the Ecological Space

Description

scatterniche displays the niche in the Ecological space (multidimensional space defined by habitat variables).

Usage

scatterniche(x, pr, xax = 1, yax = 2, pts = FALSE,
             percent = 95, clabel = 1,
             side = c("top", "bottom", "none"),
             Adensity, Udensity, Aangle, Uangle, Aborder,
             Uborder, Acol, Ucol, Alty, Ulty, Abg,
             Ubg, Ainch, Uinch, ...)

Arguments

x
a data frame giving the value of environmental variables (columns) in resource units (rows, e.g. the pixels of a raster map)

pr
a vector giving the utilisation weight for each resource unit

xax
the column number for the x-axis

yax
the column number for the y-axis

pts
logical. Whether the points should be drawn. If FALSE, minimum convex polygons are displayed

percent
100 minus the proportion of outliers to be excluded from the computation of the minimum convex polygons

clabel
a character size for the columns

side
if "top", the legend of the kept axis is upside, if "bottom" it is downside, if "none" no legend

Adensity
the density of shading lines, in lines per inch, for the available pixels polygon. See polygon for more details

Udensity
the density of shading lines, in lines per inch, for the used pixels polygon. See polygon for more details

Aangle
the slope of shading lines, given as an angle in degrees (counter-clockwise), for the available pixels polygon

Uangle
the slope of shading lines, given as an angle in degrees (counter-clockwise), for the used pixels polygon

Aborder
the color to draw the border of the available pixels polygon. See polygon for more details

Uborder
the color to draw the border of the used pixels polygon. See polygon for more details

Acol
the color for filling the available pixels polygon. if pts==FALSE, the color for the points corresponding to available pixels

Ucol
the color for filling the used pixels polygon. if pts==FALSE, the color for the points corresponding to used pixels

Alty
the line type for the available pixels polygon, as in par.

Ulty
the line type for the used pixels polygon, as in par.

Abg
if pts==TRUE, background color for open plot symbols of available pixels

Ubg
if pts==TRUE, background color for open plot symbols of used pixels

Ainch
if pts==TRUE, heigth in inches of the available pixels

Uinch
if pts==TRUE, heigth in inches of the largest used pixels

...
further arguments passed to or from other methods

Author(s)

Mathieu Basille basille@ase-research.org
Clement Calenge clement.calenge@oncfs.gouv.fr

Examples

data(chamois)
cp <- count.points(chamois$locs, chamois$map)
chamois$map
li <- kasc2df(chamois$map)
cpi <- c(cp)[li$index]
## we focus on the distance to ecotone and on the slope,
## after centring and scaling (with the help of a PCA)
scatterniche(dudi.pca(li$tab[,2:3], scannf=FALSE)$tab, cpi)
scatterniche(dudi.pca(li$tab[,2:3], scannf=FALSE)$tab, cpi, pts=TRUE)

[Package adehabitat version 1.8.18 Index]