runsNAltraj {adehabitat}	R Documentation

Highlighting the Patterns in Missing Values in Trajects

Description

runsNAltraj performs a runs test to detect any autocorrelation in the location of missing relocations, for each burst of an object of class ltraj.

summaryNAltraj returns a summary of the number and proportion of missing values for each burst of an object of class ltraj.

plotNAltraj plots the missing values in an object of class ltraj against the time.

Usage

runsNAltraj(x, nrep = 500, plotit = TRUE, ...) 

summaryNAltraj(x)

plotNAltraj(x, ...)

Arguments

x
An object of class ltraj

nrep
Number of randomisations

plotit
logical. Whether the results should be plotted on a graph

...
Further arguments to be passed to the generic function plot

Value

runsNAltraj returns a list of objects of class randtest (if a burst does not contain any missing value, the corresponding component is NULL).

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

ltraj for additional information about objects of class ltraj, setNA for additional information about missing values in such objects

Examples

## Two relocations are theoretically separated by
## 10 minutes (600 seconds)
data(puechcirc)
puechcirc

## plot the missing values
plotNAltraj(puechcirc)

## Test for an autocorrelation pattern in the missing values
(runsNAltraj(puechcirc))

[Package adehabitat version 1.8.18 Index]