simm.bb {adehabitat}	R Documentation

Brownian bridge motion

Description

This function simulates a brownian bridge motion

Usage

simm.bb(date = 1:100, begin = c(0, 0), end = begin, id = "A1", burst = id)

Arguments

date
a vector indicating the date (in seconds) at which relocations should be simulated. This vector can be of class POSIXct

begin
a vector of length 2 giving the x and y coordinates of the location beginning of the trajectory

end
a vector of length 2 giving the x and y coordinates of the location ending the trajectory

id
a character string indicating the identity of the simulated animal (see help(ltraj))

burst
a character string indicating the identity of the simulated animal (see help(ltraj))

Value

An object of class ltraj.

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr
Stephane Dray dray@biomserv.univ-lyon1.fr
Manuela Royer royer@biomserv.univ-lyon1.fr
Daniel Chessel chessel@biomserv.univ-lyon1.fr

See Also

ltraj, hbrown

Examples

plot(simm.bb(1:1000, end=c(100,100)), addpoints = FALSE)

[Package adehabitat version 1.8.18 Index]