simm.brown {adehabitat}	R Documentation

Simulate a Bivariate Brownian Motion

Description

This function simulates a Bivariate Brownian Motion.

Usage

simm.brown(date = 1:100, x0 = c(0, 0), h = 1, id = "A1", burst = id)

Arguments

date
a vector indicating the date (in seconds) at which relocations should be simulated. This vector can be of class POSIXct

x0
a vector of length 2 containing the coordinates of the startpoint of the trajectory

h
Scaling parameter for the brownian motion (larger values give smaller dispersion)

id
a character string indicating the identity of the simulated animal (see help(ltraj))

burst
a character string indicating the identity of the simulated burst (see help(ltraj))

Details

A bivariate Brownian motion can be described by a vector B2(t) = (Bx(t), By(t)), where Bx and By are unidimensional Brownian motions. Let F(t) the set of all possible realisations of the process (B2(s), 0 < s < t). F(t) therefore corresponds to the known information at time t. The properties of the bivariate Brownian motion are therefore the following: (i) B2(0)= c(0,0) (no uncertainty at time t = 0); (ii) B2(t) - B2(s) is independent of F(s) (the next increment does not depend on the present or past location); (iii) B2(t) - B2(s) follows a bivariate normal distribution with mean c(0,0) and with variance equal to (t-s).

Note that for a given parameter h, the process 1/h * B2( t * h^2 ) is a Brownian motion. The function simm.brown simulates the process B2(t * h^2). Note that the function hbrown allows the estimation of this scaling factor from data.

Value

An object of class ltraj

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr
Stephane Dray dray@biomserv.univ-lyon1.fr
Manuela Royer royer@biomserv.univ-lyon1.fr
Daniel Chessel chessel@biomserv.univ-lyon1.fr

References

~put references to the literature/web site here ~

See Also

ltraj, hbrown

Examples

plot(simm.brown(1:1000), addpoints = FALSE)

## Note the difference in dispersion:
plot(simm.brown(1:1000, h = 4), addpoints = FALSE)

[Package adehabitat version 1.8.18 Index]