simm.mou {adehabitat}	R Documentation

Simulation of a Bivariate Ornstein-Uhlenbeck Process

Description

This function simulates a bivariate Ornstein-Uhlenbeck process for animal movement.

Usage

simm.mou(date = 1:100, b = c(0, 0),
         a = diag(0.5, 2), x0 = b,
         sigma = diag(2), id = "A1",
         burst = id)

Arguments

date
a vector indicating the date (in seconds) at which relocations should be simulated. This vector can be of class POSIXct

b
a vector of length 2 containing the coordinates of the attraction point

a
a 2*2 matrix

x0
a vector of length 2 containing the coordinates of the startpoint of the trajectory

sigma
a 2*2 positive definite matrix

id
a character string indicating the identity of the simulated animal (see help(ltraj))

burst
a character string indicating the identity of the simulated burst (see help(ltraj))

Details

The Ornstein-Uhlenbeck process can be used to take into account an "attraction point" into the animal movements (Dunn and Gipson 1977). This process can be simulated using the stochastic differential equation:

dz = a (b - z(t)) dt + Sigma dB2(t)

The vector b contains the coordinates of the attraction point. The matrix a (2 rows and 2 columns) contains coefficients controlling the force of the attraction. The matrix Sigma controls the noise added to the movement (see ?simm.mba for details on this matrix).

Value

An object of class ltraj

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr
Stephane Dray dray@biomserv.univ-lyon1.fr
Manuela Royer royer@biomserv.univ-lyon1.fr
Daniel Chessel chessel@biomserv.univ-lyon1.fr

References

Dunn, J.E., & Gipson, P.S. (1977) Analysis of radio telemetry data in studies of home range. Biometrics 33: 85–101.

See Also

simm.brown, ltraj, simm.crw, simm.mba

Examples

set.seed(253)
u <- simm.mou(1:50, burst="Start at the attraction point")
v <- simm.mou(1:50, x0=c(-3,3),
              burst="Start elsewhere")
w <- simm.mou(1:50, a=diag(c(0.5,0.1)), x0=c(-3,3),
              burst="Variable attraction")
x <- simm.mou(1:50, a=diag(c(0.1,0.5)), x0=c(-3,7),
              burst="Both")
z <- c(u,v,w,x)

plot(z, addpoints = FALSE, perani = FALSE)

[Package adehabitat version 1.8.18 Index]