speed {adehabitat}	R Documentation

Computes the Speed Between Successive Relocations of an Animal - Deprecated

Description

speed measures the speed between successive relocations of animals, using objects of class traj.

Usage

speed(x, id = levels(x$id), burst = levels(x$burst), date = NULL,
      units = c("seconds", "hours", "days"))

Arguments

x
an object of class traj

id
a character vector giving the identity of the animals for which the speed is to be computed

burst
a character vector giving the identity of the circuits for which the speed is to be computed (see traj)

date
a vector of class POSIXct of length 2 (beginning, end) delimiting the period of interest

units
a character string. It determines how the speeds are computed. For example, if the coordinates are given in meters, and if units = "seconds", speeds are returned in meters per second.

Value

Returns a data frame with the following components:

id
the identity of the animal

x
if the speed is computed between the relocation 1 and 2, the x coordinate of the relocation 1.

y
if the speed is computed between the relocation 1 and 2, the y coordinate of the relocation 1.

date
a vector of class POSIXct, giving the date of relocation 1.

burst
the identity of the circuit

sp.x
the computed speed of the animal in the x direction

sp.y
the computed speed of the animal in the y direction

speed
the computed speed of the animal on the plane.

dt
the duration between the two relocations (in the units given by the parameter units).

Note

The function speed is deprecated. The class ltraj computes the speeds automatically (see ltraj).

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

traj

Examples

## Not run: 
#### Computes the speed for each wild boar
#### monitored at Puechabon
data(puechcirc)
puechcirc <- ltraj2traj(puechcirc)
puechcirc

plot(puechcirc)
sp <- speed(puechcirc)
sp[1:4,]

## End(Not run)

[Package adehabitat version 1.8.18 Index]