squirrel {adehabitat}	R Documentation

Radio-Tracking Data of Squirrels

Description

This data set describes the use and availability of 5 habitat types for 17 squirrels monitored using radio-tracking. See also the dataset squirreloc.

Usage

data(squirrel)

Format

This list has three components:

studyarea

a data frame giving the proportion of each habitat type (columns) on the study area. These proportions are repeated by rows, for all animals

mcp

a data frame giving the proportion of each habitat type (columns) in the home range of each animal (rows)

locs

a data frame giving the proportion of relocations of each animal (rows) reported in each habitat type (columns).

Source

Aebischer, N. J., Robertson, P. A. and Kenward, R. E. (1993) Compositional analysis of habitat use from animal radiotracking data. Ecology, 74, 1313–1325.

[Package adehabitat version 1.8.18 Index]