storemapattr {adehabitat}	R Documentation

Store Attributes of Maps of Class asc and kasc

Description

storemapattr stores attributes of maps of class asc and kasc in an object of class mapattr.

Usage

storemapattr(x)

Arguments

x
an object of class asc or kasc

Details

This function is essentially used by programmers in functions dealing with maps of class asc or kasc. The function getXYcoords is an example of function using storemapattr.

Value

Returns an object of class mapattr.

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

kasc and asc for additional information on the classes kasc and asc.

Examples

data(puechabon)
(kasc <- puechabon$kasc)

(toto <- storemapattr(kasc))

[Package adehabitat version 1.8.18 Index]