subsetmap {adehabitat}	R Documentation

Storing a Part of a Map

Description

subsetmap is a generic function. It has methods for the classes asc and kasc. It is used to store a part of any given map into an other object.

Usage

subsetmap(x, xlim = NULL, ylim = NULL, ...)

Arguments

x
an object of class asc or kasc

xlim
numerical vector of length 2. The x limits of the rectangle including the new map

ylim
numerical vector of length 2. The y limits of the rectangle including the new map

...
further arguments passed to or from other methods

Details

If xlim or ylim are not provided, the function asks the user to click on the map to delimit the lower left corner and the higher right corner of the new map (see Examples).

Value

Returns an object of class asc or kasc

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr, improvements by Jon Olav Vik

See Also

asc, kasc

Examples

data(puechabon)
kasc <- puechabon$kasc
el <- getkasc(kasc, "Elevation")

## limits of the new map:
xl <- c(701561, 704017)
yl <- c(3160560, 3162343)

## computation of the new map:
su <- subsetmap(el, xlim = xl, ylim = yl)

## Display
opar <- par(mar = c(0,0,0,0))
layout(matrix(c(1,1,1,1,1,1,1,1,2), byrow = TRUE, ncol = 3))
image(el, axes = FALSE)
polygon(c(xl[1], xl[2], xl[2], xl[1]),
        c(yl[1], yl[1], yl[2], yl[2]))
image(su, axes = FALSE)
box()

par(opar)
par(mfrow = c(1,1))

### Gets this part for the whole kasc object
m <- subsetmap(kasc, xlim = xl, ylim = yl)
image(m)

## Not run: 
 ## Interactive example 
 su <- subsetmap(kasc)

 image(su)

## End(Not run)

[Package adehabitat version 1.8.18 Index]