trajdyn {adehabitat}	R Documentation

Interactive Display of Objects of Class 'ltraj'

Description

This function provides an interactive version of plot.ltraj, for the exploration of objects of class ltraj.

Usage

trajdyn(x, burst = attr(x[[1]], "burst"), hscale = 1, vscale = 1,
        recycle = TRUE, display = c("guess", "windows", "tk"), ...) 

Arguments

x
an object of class ltraj

burst
a character string indicating the burst identity to explore

hscale
passed to tkrplot

vscale
passed to tkrplot

recycle
logical. Whether the trajectory should be recycled at the end of the display

display
type of display. The default guess uses a windows graphics device if getOption('device')=='windows', otherwise it uses tk (requiring the tkrplot package).

...
additional arguments to be passed to the function plot.ltraj.

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

ltraj for further information on the class ltraj, and plot.ltraj for information on arguments that can be passed to this function.

Examples

## Not run: 

## Without map
data(puechcirc)
trajdyn(puechcirc)

## With map
data(puechabon)
aa <- getkasc(puechabon$kasc, 1)
trajdyn(puechcirc, asc = aa)

## End(Not run)

[Package adehabitat version 1.8.18 Index]