vanoise {adehabitat}	R Documentation

Habitat Use by Three Species of Galliformes in the Vanoise National Parc

Description

This data frame describes the habitat use by the Black Grouse (Tetrao tetrix), the Rock Partridge (Alectoris graeca) and the Rock Ptarmigan (Lagopus mutus), in the Vanoise National Park (French Alps), from 1990 to 2000.

Usage

data(vanoise)

Format

This data frame has 3110 rows and eight columns describing the habitat composition for each occurrence of three species of Galliformes. For each located occurrence (in rows), the employees of the national park have noted: the species, the elevation (in metres), the aspect (8 classes), the habitat type (7 categories, FR means "fallen rocks") and the date (season, day, month and year).

Source

Calenge, C., Martinot, J.P. and Lebreton, P. (2003) Ecological niche separation among mountain Galliformes in the Vanoise National Parc. Game and Wildlife Science, 20, 259-285.

[Package adehabitat version 1.8.18 Index]