whale {adehabitat}	R Documentation

Argos Monitoring of Whale Movement

Description

This data set contains the relocations of one right whale.

Usage

data(whale)

Format

This data set is a regular object of class ltraj (i.e. constant time lag of 24H00)

Details

The coordinates are given in decimal degrees (Longitude - latitude).

Source

http://whale.wheelock.edu/Welcome.html

Examples

data(whale)

plot(whale)

[Package adehabitat version 1.8.18 Index]