which.ltraj {adehabitat}	R Documentation

Identify Relocations Fullfilling a Condition in an Object of Class "ltraj"

Description

This function identifies the relocations fullfilling a condition in an object of class ltraj.

Usage

which.ltraj(ltraj, expr)

Arguments

ltraj
an object of class ltraj

expr
a character string giving any syntactically correct R logical expression implying the descriptive elements in ltraj

Value

A data frame giving the ID, the Bursts and the relocations for which the condition described by expr is verified.

Author(s)

Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

ltraj for additional information about objects of class ltraj

Examples

data(puechcirc)
puechcirc

## Identifies the relocations for which time lag is
## upper than one hour
which.ltraj(puechcirc, "dt>3600")
puechcirc[burst="CH930824"][[1]][27:28,]

## Identifies the speed between successive
## relocations upper than 0.8 meters/second
which.ltraj(puechcirc, "dist/dt > 0.8")

## This is the case for example for the
## relocations #28, 58, 59 and 60 of "CH930824"
puechcirc[burst="CH930824"][[1]][c(28,58,59,60),]

[Package adehabitat version 1.8.18 Index]