niche.test {adehabitat}	R Documentation

Monte-Carlo Test on Parameters of the Ecological Niche

Description

niche.test tests for the significance of two parameters of the ecological niche of a species (marginality and tolerance), using Monte-Carlo methods. This is a bivariate test.

Usage

niche.test(kasc, points, nrep = 999, o.include = TRUE, ...) 

Arguments

kasc
a raster map of class kasc

points
a data frame with two columns, giving the coordinates of the species locations

nrep
the number of permutations

o.include
logical, passed to biv.test. If TRUE, the origin is included in the plot

...
further arguments passed to biv.test

Details

niche.test tests the significance of two parameters describing the ecological niche: the marginality (squared length of the vector linking the average available habitat conditions to the average used habitat conditions in the ecological space defined by the habitat variables), and the tolerance (inertia of the niche in the ecological space, i.e. the sum over all variables of the variance of used pixels).

At each step of the randomisation procedure, the test randomly allocates the n points in the pixels of the map. The marginality and the tolerance are then recomputed on this randomised data set.

Actual values are compared to random values with the help of the function biv.test.

Value

Returns a list containing the following components:

dfxy
a data frame with the randomized values of marginality (first column) and tolerance (second column).

obs
the actual value of marginality and tolerance.

Warning

biv.test uses the function kde2d of the package MASS.

Author(s)

Mathieu Basille basille@ase-research.org
Clement Calenge clement.calenge@oncfs.gouv.fr

See Also

biv.test for more details on bivariate tests. histniche for the histograms of the variables of the niche.

Examples

## Not run: 
data(lynxjura)

## We keep only "wild" indices.
tmp=lynxjura$loc[,4]!="D"
niche=niche.test(lynxjura$map,
                 lynxjura$locs[tmp, c("X", "Y")],
                 side = "bottom")
names(niche)

## End(Not run)

[Package adehabitat version 1.8.18 Index]