scatter.enfa {adehabitat}	R Documentation

Scatter Plot of the Results of the ENFA

Description

Performs the scatter diagrams of objects of class enfa.

Usage

## S3 method for class 'enfa'
scatter(x, xax = 1, yax = 2, pts = FALSE, nc = TRUE, 
             percent = 95, clabel = 1, side = c("top", "bottom", "none"),
             Adensity, Udensity, Aangle, Uangle, Aborder, Uborder, 
             Acol, Ucol, Alty, Ulty, Abg, Ubg, Ainch, Uinch, ...) 

Arguments

x
an object of class enfa

xax
the column number for the x-axis

yax
the column number for the y-axis

pts
logical. Whether the points should be drawn. If FALSE, minimum convex polygons are displayed

nc
whether or not the niche center should be displayed

percent
100 minus the proportion of outliers to be excluded from the computation of the minimum convex polygons

clabel
a character size for the columns

side
if "top", the legend of the kept axis is upside, if "bottom" it is downside, if "none" no legend

Adensity
the density of shading lines, in lines per inch, for the available pixels polygon. See polygon for more details

Udensity
the density of shading lines, in lines per inch, for the used pixels polygon. See polygon for more details

Aangle
the slope of shading lines, given as an angle in degrees (counter-clockwise), for the available pixels polygon

Uangle
the slope of shading lines, given as an angle in degrees (counter-clockwise), for the used pixels polygon

Aborder
the color for drawing the border of the available pixels polygon. See polygon for more details

Uborder
the color for drawing the border of the used pixels polygon. See polygon for more details

Acol
the color for filling the available pixels polygon. if pts == FALSE, the color for the points corresponding to available pixels

Ucol
the color for filling the used pixels polygon. if pts == FALSE, the color for the points corresponding to used pixels

Alty
the line type for the available pixels polygon, as in par

Ulty
the line type for the used pixels polygon, as in par

Abg
if pts == TRUE, background color for open plot symbols of available pixels

Ubg
if pts == TRUE, background color for open plot symbols of used pixels

Ainch
if pts == TRUE, heigth in inches of the available pixels

Uinch
if pts == TRUE, heigth in inches of the largest used pixels

...
further arguments passed to or from other methods

Details

scatter.enfa displays a factorial map of pixels, as well as the projection of the vectors of the canonical basis multiplied by a constant of rescaling. The kept axes for the plot are specified in a corner.

Author(s)

Mathieu Basille basille@ase-research.org

References

Basille, M., Calenge, C., Marboutin, E., Andersen, R. & Gaillard, J.M. (2008) Assessing habitat selection using multivariate statistics: Some refinements of the ecological-niche factor analysis. Ecological Modelling, 211, 233–240.

See Also

enfa, scatter

Examples

data(lynxjura)

map <- lynxjura$map

## We keep only "wild" indices.
tmp <- lynxjura$locs[,4]!="D"
locs <- lynxjura$locs[tmp, c("X","Y")]

## We perform a square root transformation
## of the variable to normalize it
map[,4] <- sqrt(map[,4])

## We perform the ENFA
tmp <- data2enfa(map, locs[tmp, c("X","Y")])
(enfa1 <- enfa(dudi.pca(tmp$tab, scannf=FALSE),
               tmp$pr, scannf = FALSE))
scatter(enfa1)

[Package adehabitat version 1.8.18 Index]